Due to limitations of file size, please reference the bootsrap files directly from https://getbootstrap.com/
This application was designed to extract saved .txt email templates for a Customer Service centre. The application assisted customer associates in wrapping up repetitive emails that initially caused slow production. 

The application uses a basic text area with a function to fetch .txt files that are saved in a directory. 

One can also set Chrome and Firefox to remember the last input text for optimization.